Feature/tests

+++ DiceGame/app/Http/Controllers/GameController.php
@@ -9,25 +9,24 @@ use Illuminate\Support\Facades\Auth;
 
 class GameController extends Controller
 {
-    public function throwDices(Request $request, int $id)
-    {
+    public function throwDices(Request $request, $id){
         // find and verify user
         $user = User::find($id);
-
-
-        if ($request->user()->id !== $user->id) {
-            return response()->json([
-                'message' => 'Unauthorized'
-            ]);
-        }
 
         if (!$user) {
             return response()->json([
                 'status' => false,
                 'message' => 'User not found',
-            ]);
+            ], 404);
         }
-
+    
+        if ($request->user()->id !== $user->id) {
+            return response()->json([
+                'status' => false,
+                'message' => 'Unauthorized'
+            ], 403);
+        }
+    
         // throw dice
         $dice1 = rand(1, 6);
         $dice2 = rand(1, 6);
@@ -44,7 +43,7 @@ class GameController extends Controller
         return response()->json([
             'message' => 'Game created successfully',
             'game' => [
-                'user_name' => $user->nickname,
+                'nickname' => $user->nickname,
                 'game' => $game,
                 'message' => $result ? 'You won!' : 'You lost.',
             ]
@@ -55,6 +54,19 @@ class GameController extends Controller
     
     // Finding user
     $user = User::find($id);
+
+    if (Auth::id() !== $user->id) {
+        return response()->json([
+            'message' => 'Unauthorized',
+        ], 403);
+    }
+
+    if (!$user) {
+        return response()->json([
+            'status' => false,
+            'message' => 'User not found',
+        ], 404);
+    }
 
     // Amount of games per player 
     $games = $user->games; 
@@ -79,11 +91,26 @@ class GameController extends Controller
         
         $user = User::find($id);
 
+        if (Auth::id() !== $user->id) {
+            return response()->json([
+                'message' => 'Unauthorized',
+            ], 403);
+        }
+
+        if (!$user) {
+            return response()->json([
+                'status' => false,
+                'message' => 'User not found',
+            ], 404);
+        }
+
         $user->games()->delete();
 
         return response()->json([
          'message' => 'Succesfully deleted'
         ]);
     }
+
+
     
 }
